refactor: extract the libs registry out of `src/app/page.tsx`
`libs` lived in a module that also imports `@/assets/*.svg`, `next/image` and `next/navigation`,
so nothing outside the Next build could read it. The MCP route got away with `import { libs } from
'@/app/page'` only because it runs inside Next, and the `browse` prototype (#579) had to duplicate
the registry outright to reach it from a plain node process.

`src/libs.ts` holds the data and its `Library` type, with no Next import. The icons leave the
registry — they are assets, not data — and `page.tsx` pairs them with the entries by key at the
point it renders them.

Closes #580
